+++ modinfo.ts
@@ -4,7 +4,7 @@ export const modinfo = defineModInfo({
   manifestVersion: 1,
   id: "irishbruse.stoneworks",
   name: "Stoneworks",
-  version: "0.0.3",
+  version: "0.0.5",
   apiVersion: 1,
   gameVersion: { minimum: "0.5.5" },
   entry: "main.js",